# smart/openapi.yaml
openapi: 3.0.3
info:
  title: PrayCalc Smart API
  version: 1.0.0
  description: |
    REST API for PrayCalc prayer time calculations, webhook notifications,
    smart home integrations (Google Home, Alexa, Home Assistant), and billing.

    **Base URL:** `https://api.praycalc.com`

    **Authentication:** Prayer times are public. Webhook management, billing,
    and smart home account linking require a Bearer JWT from Hasura Auth
    (obtained via https://auth.ummat.dev).

    **Rate Limits:** 100 requests/min for authenticated users, 30 requests/min
    for anonymous users.
  contact:
    name: PrayCalc
    url: https://praycalc.com
    email: support@praycalc.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.praycalc.com
    description: Production
  - url: http://localhost:4010
    description: Local development
paths:
...
  /api/v1/webhooks:
    post:
      operationId: createWebhook
      summary: Register a webhook
      description: |
        Register a callback URL to receive POST notifications at prayer times.
        Maximum 5 webhooks per user. The cron job fires webhooks at each
        prayer time for the registered location.
      tags:
        - Webhooks
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WebhookRegistration"
      responses:
        "201":
          description: Webhook registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WebhookResponse"
        "400":
          description: Invalid webhook data (missing callbackUrl, invalid URL, or bad coordinates)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: Maximum webhooks reached (limit 5 per user)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error: "Maximum webhooks reached"
                message: "You can register up to 5 webhooks."

    get:
      operationId: listWebhooks
      summary: List registered webhooks
      description: Returns all webhooks for the authenticated user. Users cannot see other users' webhooks.
      tags:
        - Webhooks
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of webhooks
          content:
            application/json:
              schema:
                type: object
                properties:
                  webhooks:
                    type: array
                    items:
                      $ref: "#/components/schemas/WebhookResponse"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/v1/webhooks/{id}:
    delete:
      operationId: deleteWebhook
      summary: Delete a webhook
      description: Remove a webhook registration by ID. Only the owner can delete their webhook.
      tags:
        - Webhooks
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Webhook ID
          schema:
            type: string
      responses:
        "204":
          description: Webhook deleted
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Webhook not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error: "Webhook not found"
...
  /oauth/authorize:
    get:
      operationId: oauthAuthorizeForm
      summary: OAuth authorization page
      description: |
        Renders the OAuth 2.0 consent/login page for account linking.
        Used by Google Home and Alexa during the account linking flow.
      tags:
        - OAuth
      parameters:
        - name: client_id
          in: query
          required: true
          schema:
            type: string
        - name: redirect_uri
          in: query
          required: true
          schema:
            type: string
            format: uri
        - name: state
          in: query
          required: true
          schema:
            type: string
        - name: response_type
          in: query
          required: true
          schema:
            type: string
            enum:
              - code
        - name: code_challenge
          in: query
          required: false
          schema:
            type: string
        - name: code_challenge_method
          in: query
          required: false
          schema:
            type: string
            enum:
              - S256
              - plain
      responses:
        "200":
          description: HTML consent page with login form
          content:
            text/html:
              schema:
                type: string
        "400":
          description: Unsupported response_type
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    post:
      operationId: oauthAuthorizeSubmit
      summary: Submit OAuth login
      description: |
        Processes the login form submission, authenticates via Hasura Auth,
        and redirects back to the platform with an authorization code.
        The code expires after 10 minutes and is single-use.
      tags:
        - OAuth
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              required:
                - email
                - password
                - client_id
                - redirect_uri
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                  format: password
                client_id:
                  type: string
                redirect_uri:
                  type: string
                  format: uri
                state:
                  type: string
                code_challenge:
                  type: string
      responses:
        "302":
          description: Redirect to platform with authorization code
          headers:
            Location:
              schema:
                type: string
                format: uri
              description: "redirect_uri?code=<code>&state=<state>"
        "401":
          description: Invalid credentials
          content:
            text/html:
              schema:
                type: string
        "500":
          description: Authentication service unavailable
  /oauth/token:
    post:
      operationId: oauthToken
      summary: Exchange code or refresh token
      description: |
        OAuth 2.0 token endpoint. Supports two grant types:
        - `authorization_code`: Exchange a code for access + refresh tokens.
          The code is single-use and deleted after exchange.
        - `refresh_token`: Exchange a refresh token for a new access token.
          Refresh tokens are valid for 30 days.
        Access tokens expire after 1 hour (3600 seconds).
      tags:
        - OAuth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - grant_type
              properties:
                grant_type:
                  type: string
                  enum:
                    - authorization_code
                    - refresh_token
                code:
                  type: string
                  description: Authorization code (required for authorization_code grant)
                refresh_token:
                  type: string
                  description: Refresh token (required for refresh_token grant)
                client_id:
                  type: string
      responses:
        "200":
          description: Token issued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenResponse"
        "400":
          description: Invalid or expired grant
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              examples:
                invalidGrant:
                  summary: Invalid authorization code or refresh token
                  value:
                    error: "invalid_grant"
                unsupportedGrant:
                  summary: Unsupported grant type
                  value:
                    error: "unsupported_grant_type"
  /oauth/revoke:
    post:
      operationId: oauthRevoke
      summary: Revoke a token
      description: |
        Revokes an access or refresh token. Returns 200 even if the token
        does not exist or was already revoked (RFC 7009 compliance).
      tags:
        - OAuth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                token:
                  type: string
                  description: The access or refresh token to revoke
      responses:
        "200":
          description: Token revoked (or was already invalid)
          content:
            application/json:
              schema:
                type: object
                properties:
                  revoked:
                    type: boolean
                    example: true
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Hasura Auth JWT containing `x-hasura-user-id` and `x-hasura-default-role`
        claims. Obtain from `POST https://auth.ummat.dev/signin/email-password`.
  schemas:
...
    WebhookRegistration:
      type: object
      description: Webhook registration request
      required:
        - callbackUrl
        - lat
        - lng
      properties:
        callbackUrl:
          type: string
          format: uri
          description: HTTPS URL to receive POST notifications at prayer times
          example: https://example.com/webhook
        lat:
          type: number
          format: double
          minimum: -90
          maximum: 90
          description: Latitude for prayer time calculation
          example: 40.7128
        lng:
          type: number
          format: double
          minimum: -180
          maximum: 180
          description: Longitude for prayer time calculation
          example: -74.006
        events:
          type: array
          items:
            type: string
            enum:
              - adhan
              - sunrise
              - qiyam
          description: Event types to subscribe to. Defaults to ["adhan"].
          default:
            - adhan
          example:
            - adhan
    WebhookResponse:
      type: object
      description: Registered webhook
      properties:
        id:
          type: string
          description: Unique webhook ID
        userId:
          type: string
          description: Owner user ID
        callbackUrl:
          type: string
          format: uri
          example: https://example.com/webhook
        lat:
          type: number
          format: double
          example: 40.7128
        lng:
          type: number
          format: double
          example: -74.006
        events:
          type: array
          items:
            type: string
          example:
            - adhan
        active:
          type: boolean
          example: true
        createdAt:
          type: string
          format: date-time
    TokenResponse:
      type: object
      description: OAuth token response
      properties:
        access_token:
          type: string
          description: Bearer access token (1 hour TTL)
        refresh_token:
          type: string
          description: Refresh token (30 day TTL). Only returned for authorization_code grant.
        token_type:
          type: string
          example: Bearer
        expires_in:
          type: integer
          description: Token lifetime in seconds
          example: 3600
    Error:
      type: object
      description: Error response
      properties:
        error:
          type: string
          description: Error code or short message
          example: "Invalid coordinates"
        message:
          type: string
          description: Detailed error description
          example: "lat must be -90..90, lng must be -180..180"
...